+++ CoursePlatform/CoursePlatform/urls.py
@@ -1,11 +1,16 @@
-
 from django.contrib import admin
 from django.urls import include, path
 from accounts import views
-
+from django.conf.urls.static import static
+from django.conf import settings
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include("accounts.urls")),
-    path('', include("accounts.urls"))
+    path('', include("accounts.urls")),
+    path('', include("course.urls")),
+    path('summernote/', include('django_summernote.urls')),
 ]
+
+if settings.DEBUG:
+    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)