+++ CoursePlatform/accounts/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.contrib.auth.forms import UserCreationForm
-from django_countries.fields import CountryField
-from .models import Student, Profile
+from .models import User, Profile
 from functools import partial
 
 DateInput = partial(forms.DateInput, {'class': 'datepicker'})
@@ -11,19 +10,26 @@ class FullUserCreationForm(UserCreationForm):
     """
     Extends standard usercreationform by including email address
     """
+    CHOICES = (('1', "Register as a Student",), ('2', "Register as a Partner",))
     email = forms.EmailField(required=True)
+    user_type = forms.ChoiceField(widget=forms.RadioSelect, choices=CHOICES)
 
     class Meta:
-        model = Student
-        fields = ["username", "email", "password1", "password2"]
+        model = User
+        fields = ["user_type", "username", "email", "password1", "password2"]
+        widgets = {
+            'username': forms.TextInput(attrs={'placeholder': 'Username'}),
+            'email': forms.EmailInput(attrs={'placeholder': 'E-Mail'}),
+        }
 
 
 class UserUpdateForm(forms.ModelForm):
     """
     A form for updating common fields for all types of users: email, username
     """
+
     class Meta:
-        model = Student
+        model = User
         fields = ["username", "email"]
 
 
@@ -33,7 +39,12 @@ class ProfileUpdateForm(forms.ModelForm):
     """
     profile_pic = forms.ImageField(label='Profile picture', required=False,
                                    error_messages={'invalid': "Image files only"}, widget=forms.FileInput)
+    birth_date = forms.DateField()
 
     class Meta:
         model = Profile
         fields = ['profile_pic', 'first_name', 'last_name', 'gender', 'birth_date', 'country']
+        widgets = {
+            'first_name': forms.TextInput(attrs={'placeholder': 'First Name'}),
+            'last_name': forms.TextInput(attrs={'placeholder': 'Last Name'})
+        }