+++ CoursePlatform/static/includes/contact-form.php
@@ -1,0 +1,53 @@
+
+<!----contact-section---->
+<section id="contact-section">
+  <div class="container">
+    <div class="row">
+      <h2>Contact</h2>
+      <div class="col-md-4 col-md-offset-1">
+      	<div class="contacts">
+        	<div class="icon"><img src="images/phone.png" /></div>
+            <div class="contact-info">
+            <p>Phone:(555) 123 456 789</p>
+            <p>Fax:(555)973 248</p>
+            </div>
+        </div>
+        <div class="contacts">
+        	<div class="icon"><img src="images/map.png" /></div>
+            <div class="contact-info">
+            <p>PO Box 1213</p>
+            <p>1234 Name of Location</p>
+            </div>
+        </div>
+        <div class="contacts">
+        	<div class="icon"><img src="images/letter.png" /></div>
+            <div class="contact-info">
+            <p>Email:contact@companyname.com</p>
+            </div>
+            </div>
+      </div>
+      
+      <div class="col-md-4 col-md-offset-1">
+            	<form class="contact-form">
+                	<div class="form-group">
+    <label class="sr-only" >Name</label>
+    <input type="text" class="form-control"  placeholder="Name:">
+  </div>
+  <div class="form-group">
+    <label class="sr-only">Email</label>
+    <input type="Email" class="form-control"  placeholder="Email:">
+  </div>
+  <div class="form-group">
+    <label class="sr-only">Subject</label>
+    <input type="text" class="form-control"  placeholder="Subject:">
+  </div>
+  <div class="form-group">
+  <label class="sr-only"> Message </label>
+<textarea class="form-control" rows="5" placeholder="Message:"></textarea>
+  </div>
+  <input id="submit" type="submit" value="Click Here">
+                </form>
+            </div>
+    </div>
+  </div>
+</section>